Added acceleration infrastructure

+++ src/null_engine/util/geometry/constants.hpp
@@ -2,7 +2,7 @@
 
 namespace null_engine {
 
-using FloatType = double;
+using FloatType = float;
 
 constexpr FloatType kEps = 1e-6;
